prime checker with factor checker
this adds a print line that prints the factors of the numbers if the number is prime. if it is prime there is nothing printed. 
remaining issue: this only works if the factor2 variable is a double, but this also prints a .0 at the end of the number. Find a way to remove this.