Pow 2 checker implemented
the code will now return based on how many cycles you want it to do. It will tell you the output of the machine, the step of the cycle it exited on, what cycle it is on, if the number is equivalent to a power of 2, and what that power of 2 is.